captioning_lambda/pyproject.toml
[tool.poetry]
name = "captioning_lambda"
version = "0.1.0"
description = ""
authors = ["Maciej Radzikowski <maciej@radzikowski.com.pl>"]

[tool.poetry.dependencies]
python = "~3.9"
transformers = {extras = ["torch"], version = "4.23.1"}
pillow = "9.2.0"
boto3 = "1.25.2"
aws-lambda-powertools = {extras = ["parser", "tracer"], version = "2.0.0"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
